Restructured Freestream from VL to Laplace
